login aplication

+++ list.php
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="pt-br">
+
 <head>
     <meta charset="UTF-8">
     <title>Dashboard</title>
@@ -11,20 +12,22 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
     <style>
-        .wrapper{
+        .wrapper {
             width: 600px;
             margin: 0 auto;
         }
-        table tr td:last-child{
+
+        table tr td:last-child {
             width: 120px;
         }
     </style>
     <script>
-        $(document).ready(function(){
-            $('[data-toggle="tooltip"]').tooltip();   
+        $(document).ready(function() {
+            $('[data-toggle="tooltip"]').tooltip();
         });
     </script>
 </head>
+
 <body>
     <div class="wrapper">
         <div class="container-fluid">
@@ -37,52 +40,56 @@
                     <?php
                     // Include config file
                     require_once "config.php";
-                    
-                    // Attempt select query execution
-                    $sql = "SELECT * FROM usuarios";
-                    if($result = mysqli_query($link, $sql)){
-                        if(mysqli_num_rows($result) > 0){
-                            echo '<table class="table table-bordered table-striped">';
+                    if (isset($_SESSION['email'])) {
+                        $sql = "SELECT * FROM usuarios";
+                        if ($result = mysqli_query($link, $sql)) {
+                            if (mysqli_num_rows($result) > 0) {
+                                echo '<table class="table table-bordered table-striped">';
                                 echo "<thead>";
-                                    echo "<tr>";
-                                        echo "<th>id</th>";
-                                        echo "<th>nome</th>";
-                                        echo "<th>Address</th>";
-                                        echo "<th>email</th>";
-                                        echo "<th>senha</th>";
-                                    echo "</tr>";
+                                echo "<tr>";
+                                echo "<th>id</th>";
+                                echo "<th>nome</th>";
+                                echo "<th>Address</th>";
+                                echo "<th>email</th>";
+                                echo "<th>senha</th>";
+                                echo "</tr>";
                                 echo "</thead>";
                                 echo "<tbody>";
-                                while($row = mysqli_fetch_array($result)){
+                                while ($row = mysqli_fetch_array($result)) {
                                     echo "<tr>";
-                                        echo "<td>" . $row['id'] . "</td>";
-                                        echo "<td>" . $row['nome_completo'] . "</td>";
-                                        echo "<td>" . $row['endereco'] . "</td>";
-                                        echo "<td>" . $row['email'] . "</td>";
-                                        echo "<td>";
-                                            echo '<a href="read.php?id='. $row['id'] .'" class="mr-3" title="Ver registro" data-toggle="tooltip"><span class="fa fa-eye"></span></a>';
-                                            echo '<a href="update.php?id='. $row['id'] .'" class="mr-3" title="Editar registro" data-toggle="tooltip"><span class="fa fa-pencil"></span></a>';
-                                            echo '<a href="delete.php?id='. $row['id'] .'" title="Apagar registro" data-toggle="tooltip"><span class="fa fa-trash"></span></a>';
-                                        echo "</td>";
+                                    echo "<td>" . $row['id'] . "</td>";
+                                    echo "<td>" . $row['nome_completo'] . "</td>";
+                                    echo "<td>" . $row['endereco'] . "</td>";
+                                    echo "<td>" . $row['email'] . "</td>";
+                                    echo "<td>";
+                                    echo '<a href="read.php?id=' . $row['id'] . '" class="mr-3" title="Ver registro" data-toggle="tooltip"><span class="fa fa-eye"></span></a>';
+                                    echo '<a href="update.php?id=' . $row['id'] . '" class="mr-3" title="Editar registro" data-toggle="tooltip"><span class="fa fa-pencil"></span></a>';
+                                    echo '<a href="delete.php?id=' . $row['id'] . '" title="Apagar registro" data-toggle="tooltip"><span class="fa fa-trash"></span></a>';
+                                    echo "</td>";
                                     echo "</tr>";
                                 }
-                                echo "</tbody>";                            
-                            echo "</table>";
-                            // Free result set
-                            mysqli_free_result($result);
-                        } else{
-                            echo '<div class="alert alert-danger"><em>No records were found.</em></div>';
+                                echo "</tbody>";
+                                echo "</table>";
+                                // Free result set
+                                mysqli_free_result($result);
+                            } else {
+                                echo '<div class="alert alert-danger"><em>No records were found.</em></div>';
+                            }
+                        } else {
+                            echo "Oops! Something went wrong. Please try again later.";
                         }
-                    } else{
-                        echo "Oops! Something went wrong. Please try again later.";
+                    }else{
+                        header("Location: index.php");
+                        exit();
                     }
- 
+
                     // Close connection
                     mysqli_close($link);
                     ?>
                 </div>
-            </div>        
+            </div>
         </div>
     </div>
 </body>
+
 </html>
